requirements.txt
requests
yt_dlp
easyocr
opencv-python
ffmpeg-python
ultralytics 
torch 
torchvision 
scikit-learn 
pyyaml
python-dotenv